Native refactoring
