+++ jvm_js/src/main/scala/minitest/platform/package.scala
@@ -19,19 +19,54 @@ package minitest
 
 import org.portablescala.reflect.Reflect
 
-trait Platform {
+/**
+ * The `platform` package provides the required Scala types for
+ * source-level compatibility between JVM/JS and Native, along with
+ * utilities with a platform-specific implementation.
+ */
+package object platform {
+  /**
+   * Type alias needed because Scala Native does not provide
+   * the standard [[scala.concurrent.Future]] class yet.
+   */
   type Future[+A] = scala.concurrent.Future[A]
+
+  /**
+   * Type alias needed because Scala Native does not provide
+   * the standard [[scala.concurrent.Future]] class yet.
+   */
   val Future = scala.concurrent.Future
 
+  /**
+   * Type alias needed because Scala Native does not provide
+   * an implementation for [[scala.concurrent.Await]] yet.
+   */
   val Await = scala.concurrent.Await
 
+  /**
+   * Type alias needed because Scala Native does not provide
+   * an implementation for [[scala.concurrent.Promise]] yet.
+   */
+  type Promise[A] = scala.concurrent.Promise[A]
+
+  /**
+   * Type alias needed because Scala Native does not provide
+   * an implementation for [[scala.concurrent.Promise]] yet.
+   */
   val Promise = scala.concurrent.Promise
 
+  /**
+   * Type alias needed because Scala Native does not provide
+   * an implementation for [[scala.concurrent.ExecutionContext]] yet.
+   */
   type ExecutionContext = scala.concurrent.ExecutionContext
-  val ExecutionContext = scala.concurrent.ExecutionContext
-}
 
-object Platform extends Platform {
+  /**
+   * Type alias needed because Scala Native does not provide
+   * an implementation for [[scala.concurrent.ExecutionContext]] yet.
+   */
+  val ExecutionContext = scala.concurrent.ExecutionContext
+
   type EnableReflectiveInstantiation =
     org.portablescala.reflect.annotation.EnableReflectiveInstantiation
