--- native/src/main/scala/minitest/ExecutionContext.scala
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2014-2018 by The Minitest Project Developers.
- * Some rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package minitest
-
-trait ExecutionContext
-
-object ExecutionContext {
-  val global: ExecutionContext = new ExecutionContext{}
-
-  object Implicits {
-    implicit val global: ExecutionContext = ExecutionContext.global
-  }
-}
