+++ native/src/main/scala/minitest/platform/Await.scala
@@ -16,21 +16,19 @@
  */
 
 package minitest
+package platform
 
-import scala.util.{Success, Try}
+import scala.concurrent.duration.Duration
 
-object Promise {
-  def apply[A](): Promise[A] = new Promise[A]()
-}
-
-final class Promise[A] private (private var value: Option[Try[A]] = None) {
-
-  def success(value: A): this.type = {
-    this.value = Some(Success(value))
-    this
-  }
-
-  def future: Future[A] =
-    new Future(value.getOrElse(sys.error("not completed")))
-
+/**
+  * Stub needed because Scala Native does not provide an
+  * implementation for [[scala.concurrent.Await]] yet.
+  *
+  * Note that this isn't a proper `Await` implementation,
+  * just something very simple for compilation to work and
+  * to pass the current tests.
+  */
+object Await {
+  def result[A](future: Future[A], duration: Duration): A =
+    future.value.get
 }
