+++ native/src/main/scala/minitest/platform/Promise.scala
@@ -16,10 +16,28 @@
  */
 
 package minitest
+package platform
 
-import scala.concurrent.duration.Duration
+import scala.util.{Success, Try}
 
-object Await {
-  def result[A](future: Future[A], duration: Duration): A =
-    future.value.get
+/**
+  * Stub needed because Scala Native does not provide an
+  * implementation for [[scala.concurrent.Promise]] yet.
+  *
+  * Note that this isn't a proper `Future` implementation,
+  * just something very simple for compilation to work and
+  * to pass the current tests.
+  */
+final class Promise[A] private (private var value: Option[Try[A]] = None) {
+  def success(value: A): this.type = {
+    this.value = Some(Success(value))
+    this
+  }
+
+  def future: Future[A] =
+    new Future(value.getOrElse(sys.error("not completed")))
+}
+
+object Promise {
+  def apply[A](): Promise[A] = new Promise[A]()
 }
