+++ shared/src/main/scala/minitest/TestSuite.scala
@@ -18,6 +18,7 @@
 package minitest
 
 import minitest.api._
+import minitest.platform._
 
 trait TestSuite[Env] extends AbstractTestSuite with Asserts {
   def setupSuite(): Unit = ()
