+++ shared/src/main/scala/minitest/api/package.scala
@@ -18,5 +18,5 @@
 package minitest
 
 package object api {
-  val DefaultExecutionContext = minitest.ExecutionContext.global
+  val DefaultExecutionContext = minitest.platform.ExecutionContext.global
 }
